requirements.txt
absl-py==2.1.0
asttokens==2.4.1
certifi==2024.6.2
charset-normalizer==3.3.2
comm==0.2.2
connected-components-3d==3.14.1
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.4
fonttools==4.51.0
fsspec==2024.3.1
grpcio==1.62.2
idna==3.7
imageio==2.34.0
importlib_metadata==7.1.0
ipykernel==6.29.3
ipympl==0.9.4
ipython==8.22.2
ipython-genutils==0.2.0
ipywidgets==8.1.3
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
lazy_loader==0.4
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
matplotlib-venn==0.11.10
mpmath==1.3.0
nest_asyncio==1.6.0
networkx==3.3
nibabel==5.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pip==24.0
piqa==1.3.2
platformdirs==4.2.0
pooch==1.8.2
prompt-toolkit==3.0.42
protobuf==5.26.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0
pytz==2024.1
pyvista==0.43.10
pyzmq==25.1.2
requests==2.32.3
scikit-image==0.23.1
scikit-learn==1.5.0
scipy==1.13.0
scooby==0.10.0
seaborn==0.13.2
setuptools==69.2.0
six==1.16.0
stack-data==0.6.2
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2024.4.18
torch==2.2.2
torchvision==0.17.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.2
vit-pytorch==1.6.7
vtk==9.3.0
wcwidth==0.2.13
Werkzeug==3.0.2
wheel==0.43.0
widgetsnbextension==4.0.11
zipp==3.17.0
